+++ project/plugins.sbt
@@ -1,1 +1,3 @@
 addSbtPlugin("com.codecommit" % "sbt-spiewak-sonatype" % "0.21.0")
+addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.1.0")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.6.0")
