+++ shared/src/main/scala/org/typelevel/jawn/fs2/SeqWrapper.scala
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2021 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.typelevel.jawn.fs2
+
+private[this] class SeqWrapper[A](underlying: collection.Seq[A])
+    extends collection.immutable.Seq[A] {
+  def iterator: Iterator[A] = underlying.iterator
+  def apply(i: Int): A = underlying(i)
+  def length: Int = underlying.length
+}
