+++ shared/src/main/scala/org/typelevel/jawn/fs2/package.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Typelevel
+ * Copyright 2021 Typelevel
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
