+++ shared/src/test/resources/jawnfs2/array.json
