+++ shared/src/test/resources/jawnfs2/stream.json
